backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
fpdf2==2.7.8